requirement.txt
torch>=2.0.0
torchvision
omegaconf
transformers
timm
open_clip_torch
einops
scipy
pillow
accelerate
gdown
huggingface-hub
einops
wandb
torch-fidelity
torchinfo
webdataset
accelerate==0.25.0
xformers==0.0.21
pyyaml
tqdm
loguru
numpy
pandas